fix: chain-sorting usage
