+++ src/renderer/features/proxied-add-pure/model/form-model.ts
@@ -10,7 +10,7 @@ import {
   transferableAmountBN,
   withdrawableAmountBN,
 } from '@/shared/lib/utils';
-import { sortChains } from '@/shared/lib/utils/chain-sorting';
+import { sortChains } from '@/shared/lib/utils/chains';
 import {
   createComplexTxStore,
   createInitiatorsStore,
